requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
